% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Title}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{Data frame which consists the value and the weight for each object.}

\item{W}{The knapsack size.}
}
\value{
maximum knapsack value and which elements (rows in the data.frame).
}
\description{
This function can solve the knapsack problem exact by iterating over all possible values of W.
}
